shell: move shell_setup.py to subsystems
