+++ src/python/pants/backend/shell/goals/tailor.py
@@ -7,7 +7,7 @@ import os
 from dataclasses import dataclass
 from typing import Iterable
 
-from pants.backend.shell.shell_setup import ShellSetup
+from pants.backend.shell.subsystems.shell_setup import ShellSetup
 from pants.backend.shell.target_types import (
     ShellSourcesGeneratorTarget,
     Shunit2TestsGeneratorSourcesField,
