+++ src/python/pants/backend/shell/subsystems/shell_setup.py
