+++ src/python/pants/backend/shell/util_rules/shell_command.py
@@ -10,7 +10,7 @@ import shlex
 from dataclasses import dataclass
 from textwrap import dedent  # noqa: PNT20
 
-from pants.backend.shell.shell_setup import ShellSetup
+from pants.backend.shell.subsystems.shell_setup import ShellSetup
 from pants.backend.shell.target_types import (
     ShellCommandCommandField,
     ShellCommandExtraEnvVarsField,
